style: move LanguageHandler and related logic to separate file
